# Introduce buzhash chunker      It has the same properties as Rabin but is much faster. Benchmark results of just chunkers: