```
name             speed
Buzhash2/1K-4    1.71GB/s ± 4%
Buzhash2/1M-4    1.15GB/s ± 2%
Buzhash2/16M-4    892MB/s ± 1%
Buzhash2/100M-4   904MB/s ± 2%
Rabin/1K-4       13.8MB/s ± 3%
Rabin/1M-4        171MB/s ± 3%
Rabin/16M-4       182MB/s ± 4%
Rabin/100M-4      182MB/s ± 3%
Default/1K-4     1.74GB/s ± 4%
Default/1M-4     3.22GB/s ± 2%
Default/16M-4    3.88GB/s ± 2%
Default/100M-4   4.21GB/s ± 6%
```